FirstTime Commit ver 1.0
This is first time I'm publishing this tool. The development of this
tool was made during my work in eToro. Therefore all credentials saved
to eToro. I removed all confidential data from the code, but still there
are  some  not confidential leftovers like eToro icons or old tool name
(a.k.a eToroAutoStudio).